backend/main.py
from fastapi import FastAPI, UploadFile, File
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel

app = FastAPI()

# 1. The Anti-Blocker: Allow Next.js to talk to FastAPI
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000"], # Your frontend URL
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# 2. Data Schemas
class TriageRequest(BaseModel):
    media_url: str
    equipment_model: str

# 3. Step 1: The Initial Audio/Video Diagnostic
@app.post("/api/triage")
async def initial_triage(request: TriageRequest):
    """
    Person 2's Gemini script gets plugged in here.
    Your ElevenLabs script gets plugged in here.
    """
    # TODO: Pass request.media_url to Gemini Multimodal
    # TODO: Pass Gemini's text output to ElevenLabs
    
    return {
        "status": "success",
        "fault": "Faulty Igniter", 
        "next_action": "Remove the lower front panel to access the burner compartment.",
        "audio_url": "https://api.elevenlabs.io/.../mock_audio.mp3" # Placeholder
    }

# 4. Step 2: The Visual Compositing Fallback
@app.post("/api/visual-assist")
async def visual_assist(image: UploadFile = File(...)):
    """
    Person 2's spatial Gemini script gets plugged in here.
    Person 3's Cloudinary overlay script gets plugged in here.
    """
    # TODO: Send new image to Gemini for X/Y coordinates
    # TODO: Pass coordinates to Cloudinary URL builder
    
    return {
        "status": "success",
        "composite_url": "https://res.cloudinary.com/.../mock_overlay.jpg" # Placeholder
    }

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)