backend/requirements.txt
# Web Server & Framework
fastapi
uvicorn
python-multipart    # Required for handling image/video file uploads in FastAPI

# Environment Management
python-dotenv       # For loading your API keys from a .env file

# AI & Media APIs
google-genai # For Gemini Multimodal RAG
cloudinary          # For Person 3's visual compositing
elevenlabs          # For your voiceover integration
requests            # Standard library for any webhook/API fallbacks